send media messages, upload to cloud

+++ src/main/java/com/myanime/infrastructure/entities/Message.java
@@ -19,7 +19,7 @@ public class Message {
 
     private String senderId;
 
-    private Byte messageType;
+    private Short messageType;
 
     private String content;
